requirements.txt
# Core Dependencies
Flask==3.1.1
Flask-CORS==6.0.1
python-dotenv==1.1.0
Werkzeug==3.1.3
gunicorn==23.0.0

# AI and Machine Learning
torch==2.7.0
transformers==4.52.4
sentence-transformers==4.1.0
faiss-cpu==1.11.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
einops==0.8.1
accelerate==1.7.0
bitsandbytes==0.46.0
unsloth==2025.6.3
unsloth-zoo==2025.6.2
trl==0.19.0
peft==0.15.2

# Audio Processing
webrtcvad==2.0.10
SpeechRecognition==3.14.3
pydub==0.25.1
happytransformer==2.4.1

# Location and Geocoding
geopy==2.4.1
geocoder==1.38.1

# Security and Rate Limiting
better-profanity==0.7.0
bleach==6.2.0
Flask-Limiter==3.12

# Cloud Services and Tunneling
cloudinary==1.44.1
pyngrok==7.2.11

# Utilities
python-dateutil==2.9.0.post0
requests==2.32.3
tqdm==4.67.1
pydantic==1.10.9
typing-extensions==4.13.2
tzdata==2024.2

# NLP and Text Processing
spacy==3.7.2
wasabi==1.1.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Additional Dependencies
python-Levenshtein==0.27.1
fuzzywuzzy==0.18.0
urllib3==2.4.0
websockets==13.0
redis==5.2.1
dataclasses-json==0.6.4
uuid==1.30

# Additional Dependencies for Multi-User Support
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==42.0.5

# Advanced Processing Dependencies
psutil==6.1.0
numpy==1.26.4
scipy==1.13.0

# Error Recovery and Monitoring
prometheus-client==0.20.0
sentry-sdk==2.17.0